certsuite pod and sidecar renamed to remove the cnf- prefix
Also minor updates in README.md
